Fixed bug in state transition logic. Simplified the state logic, just go straight to using the external crystal.
Calibration completes in about 3 seconds currently.
This might be able to happen faster, but it takes a bunch of time to test and there's deminishing returns. 3 seconds is plenty fast enough for pre-match setup. The robot can be moving during initialization and calibration, so no worrries about introducing error if the robot is turned on too early (like with a rate gyro).

Added some additional diagnostics methods so that you can tell what's going on with the sensor from outside the class. It should transition through these states:
  Establish communications
  Initialization
  Calibration

Upped the internal thread period to 20ms, might be able to go a little faster (10ms would be the fastest the sensor could provide data in some modes).

Closes #1 and #4.

Changed default mode of operation to IMU, in this mode the magnetometer isn't used, relative position heading is provided as a result. So no heading jumps will occur as the sensor completes calibration sequence. Closes #3